Rename net package and drop prefix for most classes
