+++ example/src/main/java/org/geysermc/mcprotocollib/protocol/example/MinecraftProtocolTest.java
@@ -21,8 +21,8 @@ import org.geysermc.mcprotocollib.network.event.server.SessionAddedEvent;
 import org.geysermc.mcprotocollib.network.event.server.SessionRemovedEvent;
 import org.geysermc.mcprotocollib.network.event.session.DisconnectedEvent;
 import org.geysermc.mcprotocollib.network.event.session.SessionAdapter;
-import org.geysermc.mcprotocollib.network.net.NetClientSession;
-import org.geysermc.mcprotocollib.network.net.NetServer;
+import org.geysermc.mcprotocollib.network.netty.NetClientSession;
+import org.geysermc.mcprotocollib.network.netty.NetServer;
 import org.geysermc.mcprotocollib.network.packet.Packet;
 import org.geysermc.mcprotocollib.protocol.MinecraftConstants;
 import org.geysermc.mcprotocollib.protocol.MinecraftProtocol;
