+++ protocol/src/main/java/org/geysermc/mcprotocollib/network/netty/BundlerUnpackerDecoder.java
@@ -1,4 +1,4 @@
-package org.geysermc.mcprotocollib.network.net;
+package org.geysermc.mcprotocollib.network.netty;
 
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.MessageToMessageDecoder;
@@ -9,7 +9,7 @@ import org.geysermc.mcprotocollib.protocol.packet.ingame.clientbound.Clientbound
 import java.util.ArrayList;
 import java.util.List;
 
-public class NetBundlerUnpacker extends MessageToMessageDecoder<MinecraftPacket> {
+public class BundlerUnpackerDecoder extends MessageToMessageDecoder<MinecraftPacket> {
     private List<MinecraftPacket> currentPackets;
 
     @Override
