+++ protocol/src/main/java/org/geysermc/mcprotocollib/network/netty/PacketSizerCodec.java
@@ -1,4 +1,4 @@
-package org.geysermc.mcprotocollib.network.net;
+package org.geysermc.mcprotocollib.network.netty;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
@@ -12,7 +12,7 @@ import org.geysermc.mcprotocollib.network.packet.PacketHeader;
 import java.util.List;
 
 @RequiredArgsConstructor
-public class NetPacketSizer extends ByteToMessageCodec<ByteBuf> {
+public class PacketSizerCodec extends ByteToMessageCodec<ByteBuf> {
     private final PacketHeader header;
     private final PacketCodecHelper codecHelper;
