+++ protocol/src/test/java/org/geysermc/mcprotocollib/protocol/MinecraftProtocolTest.java
@@ -7,8 +7,8 @@ import org.geysermc.mcprotocollib.network.Server;
 import org.geysermc.mcprotocollib.network.Session;
 import org.geysermc.mcprotocollib.network.event.session.DisconnectedEvent;
 import org.geysermc.mcprotocollib.network.event.session.SessionAdapter;
-import org.geysermc.mcprotocollib.network.net.NetClientSession;
-import org.geysermc.mcprotocollib.network.net.NetServer;
+import org.geysermc.mcprotocollib.network.netty.NetClientSession;
+import org.geysermc.mcprotocollib.network.netty.NetServer;
 import org.geysermc.mcprotocollib.network.packet.Packet;
 import org.geysermc.mcprotocollib.protocol.codec.MinecraftCodec;
 import org.geysermc.mcprotocollib.protocol.data.game.entity.player.GameMode;
